Propose "why" folder that we will keep as "advisors" knowledge base
I would like to propose the "why" folder, where we will keep a
(very) short summary about some of the rules and policies that
we will be - as Advisors - discussing with the projects we will
advise.

I think we need something like that - also as a place where we
- advisors - will be able to update and exchange our understanding
on why there are certain rules and policies. Also to discuss future
variations and changes to those rules, possibly resulting from
interactions with multiple PMCs when we will learn from them how
they are applying certain rules and policies in their environment.

I gave an example of how I see such "why" captured information could
look like - starting with the most familiar (for me) release process
and I am curious what others think about it.