Add bi-directional approach for Advisors
This is something I wanted to proposed from the beginnig of seing
the Sharpenes apporach. I believe the original approach where
Advisors just advise the projects their join - and the flow is in
one direction, they should treat the opportunity of interacting
with different PMCs as an opportunity to learn from the projects
and discuss with other advisors whether the current incarnation
of the rules we have is applicable in practice and whether maybe
they should be adjusted and improved or maybe they should be
updated to be better applicable to different projects - coming from
various culturs, backgrounds, projects with different sizes,
with different stakeholder engagement.

Having the relation works in both directions is an opportunity
to learn and adapt, and to exchange findings between the Advisors
that might lead to better adaptability and better common understanding
of the rules and policies we have, and more participatory role of
both Advisors and PMCs that might get beter chance to shape the
way how Foundation works on an ongoing way, as well as to give
the board an opportunity to listen to the voices and better understand
the varuious needs of various projects.