Add FAQ on how to start (and end) Advisor relationship with PMC
This PR adds quite a bit of clarity on how the process of initiating
(and ending) the relationship between the Advispr and PMC is expected
to happen. It's more on a FAQ rather than rules, and it describes the
usual way it can happen, with focus put on the voluntary nature of
being an Advisor and clarifies the role of Board in the process.

Co-authored-by: tison <wander4096@gmail.com>