Add a chapter about "collaborative" aspect of Advisors
Having a - potentially - bigger and engaged group of Advisors that
can exchange best practices they learn from the project and have
the possibility of shaping best practices that other Advisors will
also disseminate among other projects, is probably one of the
biggest values of a group such as Advisors and we should state it
explicitly that such collaboration is expected and encouraged.

Also this might be one of the important incentives for Advisors to
get engaged, and for PMCs to ask for Advisors when they know that
whatever they do with Advisors is not only to apply "top-bottom"
rules and principles, but also to be able to share their own
practices and ideas and variations coming from different stages,
size, stakeholder engagement they are in - and knowing that they
can also shape the common understanding and interpretation of the
rules and principles the ASF is built on.

Fingers crossed that we will be able to incentivise enough engaged
members and PMCs to make Advisors a useful platformm for this kind
of collaboration.