Propose "why" explanation on discussions and decision making
This change is a proposal on how we could describe the next "why" that
I consider as very important - discussions and decision making.

I tried to map all the red-flags we had for it into a coherent
description of the "why" and linked to the two important places
in the ASF "rules" that best describe some of the aspects of it.

I think the part of the variations might be contentious, and I hope
that in the advisors "leaders" group we can discuss it and get to
a consensus that

a) there are good reasons why the mailing lists were chosen in the
   long past

b) the reasons for that are still valid today and mailing list shoud
   remoain the most important place for discussions and decision making

while c) with the current set of technologies in many cases it is not
the only way, but as long as PMC actively makes sure that the important
discussions and decisions (and generally only those in more active
projects) are brought to the list and where list is active and used
for all the voting, super important decisions, this is all cool

Looking forward to discussion here.